PostDeployment with test tasks and small fix

+++ HistoryQuest.DataBase/dbo/Tables/Tasks.sql
@@ -2,7 +2,7 @@
     [id]             BIGINT           IDENTITY (1, 1) NOT NULL,
     [gid]            UNIQUEIDENTIFIER NOT NULL,
     [Text]           TEXT             NOT NULL,
-    [MaxScore]       BIGINT           NOT NULL,
+    [MaxScore]       INT              NOT NULL,
     [TaskTypeGID]    UNIQUEIDENTIFIER NOT NULL,
     [CheckPointGID]  UNIQUEIDENTIFIER NOT NULL,
     [SourceFilePath] NVARCHAR (250)   NOT NULL,
@@ -12,6 +12,8 @@
     CONSTRAINT [FK_Tasks_TaskTypes] FOREIGN KEY ([TaskTypeGID]) REFERENCES [dbo].[TaskTypes] ([gid]),
     CONSTRAINT [FK_Tasks_Users] FOREIGN KEY ([AuthorGID]) REFERENCES [dbo].[Users] ([gid])
 );
+
+
